// src/entities/work/api/fetchPlannedTasks.generated.ts
export type FetchPlannedTasksQuery = {
  __typename?: 'query_root';
  public7_planned_tasks: Array<{
    __typename?: 'public7_planned_tasks';
    id: any;
    name: string;
    description?: string | null;
    rm_task_id: any;
    yaml_url?: string | null;
    time_work_id: any;
    author_id: any;
  }>;
};